+++ run_celery.sh
@@ -1,0 +1,1 @@
+celery -A config worker -l info